# docs/openapi-analysis.yaml
openapi: 3.1.0
info:
  title: Ruflo Analysis API
  version: 3.0.0
  description: Workspace-scoped analysis API. All endpoints require auth and workspace membership.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /analysis/jobs:
    get:
      tags: [analysis]
      summary: List analysis jobs
      operationId: listAnalysisJobs
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AnalysisJobStatus'
      responses:
        '200':
          description: Analysis jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [analysis]
      summary: Create analysis job
      operationId: createAnalysisJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnalysisJobRequest'
      responses:
        '201':
          description: Created analysis job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /analysis/jobs/{job_id}:
    get:
      tags: [analysis]
      summary: Get analysis job
      operationId: getAnalysisJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Analysis job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /analysis/jobs/{job_id}/compare:
    post:
      tags: [analysis]
      summary: Compare an analysis job with documents
      operationId: compareAnalysisJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareRequest'
      responses:
        '202':
          description: Job with comparison result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /analysis/jobs/{job_id}/summarize:
    post:
      tags: [analysis]
      summary: Run deterministic summary generation
      operationId: summarizeAnalysisJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SummarizeRequest'
      responses:
        '202':
          description: Job with analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /analysis/jobs/{job_id}/approve:
    post:
      tags: [analysis]
      summary: Approve completed analysis job
      operationId: approveAnalysisJob
      description: Only completed jobs can be approved. Approval is idempotent and only updates suggestion status.
      parameters:
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRequest'
      responses:
        '200':
          description: Approved job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /analysis/jobs/{job_id}/result:
    get:
      tags: [analysis]
      summary: Get analysis result
      operationId: getAnalysisResult
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /analysis/results/{result_id}:
    get:
      tags: [analysis]
      summary: Get analysis result by result ID
      operationId: getAnalysisResultById
      parameters:
        - $ref: '#/components/parameters/ResultId'
      responses:
        '200':
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResultDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /analysis/results/{result_id}/review:
    post:
      tags: [analysis]
      summary: Mark result for review (draft → review)
      operationId: markAnalysisResultForReview
      description: Submits a draft result for admin review. Requires workspace membership.
      parameters:
        - $ref: '#/components/parameters/ResultId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: Result in review status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResultDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /analysis/results/{result_id}/approve:
    post:
      tags: [analysis]
      summary: Approve analysis result (review → approved)
      operationId: approveAnalysisResult
      description: |
        Requires workspace_admin role. `confirm: true` is mandatory (PROHIBIT-08).
        8-rule AnalysisApprovalPolicy is enforced server-side.
        Sets `approved_by` and `approved_at`.
      parameters:
        - $ref: '#/components/parameters/ResultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveAnalysisResultRequest'
      responses:
        '200':
          description: Approved result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResultDetail'
        '400':
          description: confirm is false or missing
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /analysis/results/{result_id}/reject:
    post:
      tags: [analysis]
      summary: Reject analysis result (review → rejected)
      operationId: rejectAnalysisResult
      description: Requires workspace_admin role. `reason` is mandatory.
      parameters:
        - $ref: '#/components/parameters/ResultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectAnalysisResultRequest'
      responses:
        '200':
          description: Rejected result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResultDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /analysis/results/{result_id}/import:
    post:
      tags: [analysis]
      summary: Import approved analysis result into knowledge base
      operationId: importAnalysisResult
      description: |
        Requires workspace_admin role and `result.status = 'approved'` (HTTP 409 otherwise).
        Idempotent: repeated calls produce created=0, found=N — no duplicates.
        KB effects: tags (find-or-create), document_tags (upsert, source='ki'),
        topics (slug-based find-or-create, status='draft'), topic_documents, topic_tags.
        PROHIBIT-08: topics are always created with status='draft', never auto-approved.
      parameters:
        - $ref: '#/components/parameters/ResultId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: Import statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportAnalysisResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Result is not in approved status
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
    ResultId:
      name: result_id
      in: path
      required: true
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    Unauthorized:
      description: Missing or invalid authentication
    Forbidden:
      description: Missing workspace permission or insufficient role
    NotFound:
      description: Resource not found in workspace
    Conflict:
      description: Invalid status transition or status guard violation
    ValidationError:
      description: Invalid request payload
  schemas:
    AnalysisJobStatus:
      type: string
      enum: [queued, pending, running, completed, failed, cancelled, approved, rejected]
    AnalysisResultStatus:
      type: string
      enum: [draft, review, approved, rejected]
...
    ApprovalDecision:
      type: string
      enum: [approved]
    AnalysisResult:
      type: object
      required: [id, job_id, summary, key_points, suggested_tags, suggested_topics, confidence, created_at]
      properties:
        id:
          type: string
        job_id:
          type: string
        summary:
          type: string
        key_points:
          type: array
          items:
            type: string
        suggested_tags:
          type: array
          items:
            type: string
        suggested_topics:
          type: array
          items:
            type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        created_at:
          type: string
          format: date-time
...
    AnalysisJobListItem:
      type: object
      required:
        - id
        - workspace_id
        - source_document_ids
        - status
        - analysis_type
        - prompt
        - created_by
        - created_at
        - started_at
        - finished_at
        - error_code
        - error_message
      properties:
        id:
          type: string
        workspace_id:
          type: string
        source_document_ids:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/AnalysisJobStatus'
        analysis_type:
          type: string
        prompt:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        error_code:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
    AnalysisJobResponse:
      allOf:
        - $ref: '#/components/schemas/AnalysisJobListItem'
        - type: object
          required: [result, comparison, suggestions]
          properties:
            result:
              $ref: '#/components/schemas/AnalysisResult'
              nullable: true
            comparison:
              $ref: '#/components/schemas/AnalysisComparison'
              nullable: true
            suggestions:
              type: array
              items:
                $ref: '#/components/schemas/AnalysisSuggestion'
    AnalysisJobListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AnalysisJobListItem'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
    CreateAnalysisJobRequest:
      type: object
      additionalProperties: false
      required: [source_document_ids, analysis_type, prompt]
      properties:
        source_document_ids:
          type: array
          minItems: 1
          items:
            type: string
        analysis_type:
          type: string
          minLength: 1
          maxLength: 64
        prompt:
          type: string
          minLength: 1
    CompareRequest:
      type: object
      additionalProperties: false
      properties:
        compared_document_ids:
          type: array
          items:
            type: string
          default: []
        max_differences:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
    SummarizeRequest:
      type: object
      additionalProperties: false
      properties:
        prompt:
          type: string
          minLength: 1
          nullable: true
        max_suggestions:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
    ApproveRequest:
      type: object
      additionalProperties: false
      properties:
        decision:
          $ref: '#/components/schemas/ApprovalDecision'
          default: approved
    AnalysisResultDetail:
      type: object
      required: [id, job_id, status, summary, suggested_tags, suggested_topics, key_points, confidence, created_at]
      properties:
        id:
          type: string
        job_id:
          type: string
        status:
          $ref: '#/components/schemas/AnalysisResultStatus'
        summary:
          type: string
        key_points:
          type: array
          items:
            type: string
        suggested_tags:
          type: array
          items:
            type: string
        suggested_topics:
          type: array
          items:
            type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        approved_by:
          type: string
          nullable: true
        approved_at:
          type: string
          format: date-time
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    ApproveAnalysisResultRequest:
      type: object
      required: [confirm]
      additionalProperties: false
      properties:
        confirm:
          type: boolean
          description: Must be true. Rejects with 400 if false or missing. (PROHIBIT-08)
        reviewer_note:
          type: string
          nullable: true
    RejectAnalysisResultRequest:
      type: object
      required: [reason]
      additionalProperties: false
      properties:
        reason:
          type: string
          minLength: 1
    ImportAnalysisResultResponse:
      type: object
      required:
        - result_id
        - tags_created
        - tags_found
        - document_tags_applied
        - topics_created
        - topics_found
        - topic_docs_attached
        - topic_tags_applied
        - source_document_count
      additionalProperties: false
      properties:
        result_id:
          type: string
          description: UUID des importierten AnalysisResult
        tags_created:
          type: integer
          minimum: 0
          description: Neu angelegte Tags
        tags_found:
          type: integer
          minimum: 0
          description: Bereits existierende Tags (unverändert)
        document_tags_applied:
          type: integer
          minimum: 0
          description: Eingetragene document_tags (source='ki')
        topics_created:
          type: integer
          minimum: 0
          description: Neu angelegte Topics (status='draft')
        topics_found:
          type: integer
          minimum: 0
          description: Bereits existierende Topics
        topic_docs_attached:
          type: integer
          minimum: 0
          description: Verknüpfte topic_documents
        topic_tags_applied:
          type: integer
          minimum: 0
          description: Verknüpfte topic_tags
        source_document_count:
          type: integer
          minimum: 0
          description: Anzahl Quelldokumente aus dem AnalysisJob